ai-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
openai==1.107.2
replicate==1.0.7
python-dotenv==1.1.1
